Feat/sj/share todo room

+++ DDIS_Project/src/main/java/com/DDIS/shareTodo/Command/domain/aggregate/Entity/MemberShareTodo.java
@@ -15,11 +15,12 @@ public class MemberShareTodo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "member_share_todo_num")
-    private int memberShareTodoNum;
+    private Long memberShareTodoNum;
 
     @ManyToOne
     @JoinColumn(name = "share_todo_num")
     private ShareTodo shareTodoNum;
+
     @ManyToOne
     @JoinColumn(name = "member_num")
     private Members memberNum;
